+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.6.1] - 2025-12-10
+
+### Changed
+
+- Added new `custombrowserstructs.CustomBrowserDefinition`
+- Added new `HideConditionOperand` values for PayloadType's BuildParameter Hide Conditions
+- Added `VerifierRegex` option to Command Parameters
+- Updated `SendMythicRPCCallbackCreate` to allow supplying `Cwd` and `ImpersonationContext` parameters
+- Updated `SendMythicRPCCallbackUpdate` to allow supplying `Cwd`, `ImpersonationContext`, and `Dead` parameters
+- Added `SendMythicRPCCallbackTokenSearch`
+- Added `SendMythicRPCCustomBrowserSearch`
+- Added new Service `MythicServiceCustomBrowser` when starting your main program
+- Fixed `Webhook` and `Logger` services not responding to `OnContainerStart` messages
+- Updated `ContainerVersion` to `"v1.4.1"`
+
 ## [v1.6.0] - 2025-10-29
 
 ### Changed